+++ locales/enUS.lua
@@ -29,6 +29,17 @@ L['PLAYER_ALREADY_IN_GUILD'] = "is already in a guild"
 L['PLAYER_JOINED_GUILD'] = "has joined the guild"
 L['PLAYER_DECLINED_INVITE'] = "declines your guild invitation"
 
+--? 3.2 Changes
+L['SYNC_ALREADY_IN_PROGRESS'] = 'Sync already in progress with'
+L['NO_CLIENTS_FOUND'] = 'No clients found to sync with.'
+L['NO_BLACKLISTED_ADDED'] = 'No blacklisted players added.'
+L['NO_ANTISPAM_ADDED'] = 'No players added to the Anti-Spam list.'
+L['CLIENT_REQUEST_DATA_TIMEOUT'] = 'Data request timeout for'
+L['FAILED_TO_RECEIVE_SYNC_DATA'] = 'Failed to receive sync data from'
+L['OUTDATED_VERSION'] = 'is using an outdated version'
+L['OLDER_VERSION'] = 'is using an older version'
+L['NEWER_VERSION'] = 'is using a newer version'
+
 -- * Donation Message
 L['DONATION_MESSAGE'] = [[
 I hope you find this addon useful. I have put a lot of time and effort into
@@ -43,7 +54,7 @@ L['ABOUT_TOOLTIP'] = 'See What\'s New and support info.'
 L['PLEASE_WAIT'] = 'Please wait'
 L['ERROR_SCAN_WAIT'] = 'seconds before scanning again.'
 
--- * Version 3.0 Changes
+--? Version 3.0 Changes
 
 -- * GR Icon Bar
 L['LOCK'] = 'Lock'
@@ -54,12 +65,12 @@ L['SYNC'] = 'Sync'
 L['SYNC_TOOLTIP'] = 'Manual sync with guild members.'
 L['ANALYTICS'] = 'Analytics'
 L['ANALYTICS_TOOLTIP'] = 'View the analytics window.'
-L['BLACKLIST'] = 'Black List'
+L['BLACKLIST'] = 'Blacklist'
 L['NO_REASON'] = 'No reason given.'
 L['BLACKLIST_TOOLTIP'] = 'Add a player to the blacklist.'
-L['BLACK_LIST_REASON_INPUT'] = 'Why do you want to black list?'
+L['BLACK_LIST_REASON_INPUT'] = 'Why do you want to blacklist?'
 L['BL_NAME_NOT_ADDED'] = 'Black list name not added.'
-L['ADDED_TO_BLACK_LIST'] = 'was added to the black list with %s as a reason.'
+L['ADDED_TO_BLACK_LIST'] = 'was added to the blacklist with %s as a reason.'
 L['FILTER_EDITOR'] = 'Filter Editor'
 L['FILTER_EDITOR_TOOLTIP'] = 'Open the filter editor window.'
 L['COMPACT_MODE'] = 'Compact Mode'
@@ -86,7 +97,7 @@ L['GUILDNAME'] = 'GUILDNAME'
 L['NO_GUILD_NAME'] = 'No Guild Name'
 L['PLAYERNAME'] = 'PLAYERNAME'
 L['NO_PLAYER_NAME'] = 'player'
-L['BLACK_LIST'] = 'Black List'
+L['BLACK_LIST'] = 'Blacklist'
 L['OK_INVITE'] = 'Do you want to invite anyway?'
 L['INVITE_REJECTED'] = 'Message not sent, appears player has Block Guild Invites on.'
 L['NO_INVITE_MESSAGE'] = 'No message selected. Please select a message in the home screen.'
@@ -113,11 +124,11 @@ L['RACES'] = 'Races'
 -- * GR Scanner Messages
 L['BL'] = 'BL'
 L['BLACK_LISTED'] = 'Black listed'
-L['IS_ON_BLACK_LIST'] = 'is on the Black List.'
-L['BLACK_LIST_CONFIRM'] = 'Are you sure you want to add this player to the black list?'
+L['IS_ON_BLACK_LIST'] = 'is on the Blacklist.'
+L['BLACK_LIST_CONFIRM'] = 'Are you sure you want to add this player to the blacklist?'
 L['ANTI_SPAM'] = 'Anti-Spam'
 L['READY_INVITE'] = 'Ready to invite'
-L['BL_ADD_PLAYER'] = 'Add Player to black list'
+L['BL_ADD_PLAYER'] = 'Add Player to blacklist'
 L['SKIP'] = 'Skip'
 L['SKIP_DESC'] = 'Skip the current player and move to the next player.'
 L['WHO_RESULTS'] = 'Who Results'
@@ -127,7 +138,7 @@ L['RESETTING_FILTERS'] = 'Resetting filters on next scan.'
 L['NUMBER_PLAYERS_FOUND'] = 'Number of players found'
 L['INVITE_BUTTON_TOOLTIP'] = 'Invite player to the guild.'
 L['INVITE_BUTTON_BODY_TOOLTIP'] = [[Only unchecked players will be invited.
-Checked players are for black list and skip.]]
+Checked players are for blacklist and skip.]]
 L['BL_BUTTON_TOOLTIP'] = 'Add player to the blacklist.'
 L['BL_BUTTON_BODY_TOOLTIP'] = [[Add the player to the blacklist and skip to the next player.
 Can only invite if no players are checked.]]
@@ -143,7 +154,7 @@ L['TOTAL_INVITED'] = 'Players Invited'
 L['INVITES_PENDING'] = 'Invites Pending'
 L['TOTAL_DECLINED'] = 'Invites Declined'
 L['TOTAL_ACCEPTED'] = 'Invites Accepted'
-L['TOTAL_BLACKLISTED'] = 'Black Listed'
+L['TOTAL_BLACKLISTED'] = 'Blacklisted'
 L['TOTAL_ANTI_SPAM'] = 'Players on Anti-Spam'
 L['SESSION_STATS'] = 'Session Stats'
 
@@ -251,8 +262,8 @@ L['MAX_CHARS'] = '(<sub> characters per message)'
 L['LENGTH_INFO'] = 'Assumes 12 characters when using PLAYERNAME'
 L['MESSAGE_LENGTH'] = 'Message Length'
 L['BL_PRIVATE_REASON'] = 'Toggle Private Reason'
-L['BL_PRIVATE_REASON_DESC'] = 'Toggle the private reason for black listing.'
-L['BL_PRIVATE_REASON_ERROR'] = 'You did not black list'
+L['BL_PRIVATE_REASON_DESC'] = 'Toggle the private reason for blacklisting.'
+L['BL_PRIVATE_REASON_ERROR'] = 'You did not blacklist'
 
 -- * GR GM Invite Settings Window
 L['GM_INVITE'] = 'GM Messages'
@@ -293,9 +304,9 @@ L['INVITE_MESSAGES_DESC'] = [[These messages are separate from the GM synced mes
 They are tied only to your guild characters.]]
 
 -- * BlackList Settings Window
-L['BLACK_LIST'] = 'Black List'
-L['BLACK_LIST_REMOVE'] = 'Remove Selected Black List Entries'
-L['ADD_TO_BLACK_LIST'] = 'Add player to black list.'
+L['BLACK_LIST'] = 'Blacklist'
+L['BLACK_LIST_REMOVE'] = 'Remove Selected Blacklist Entries'
+L['ADD_TO_BLACK_LIST'] = 'Add player to blacklist.'
 
 -- * Invalid Settings Window
 L['INVALID_ZONE'] = 'Invalid Zones'