+++ shared/GR_Dataset.lua
@@ -49,8 +49,11 @@ function ds:WhatsNew() -- What's new in the current version
                     Discord: https://discord.gg/ZtS6Q2sKRH
                 (or click on the icon in the top left corner)|r
 
-    |CFFFFFF00v3.2.41 Notes|r
-        - Working on sync.
+    |CFFFFFF00v3.2.43 Notes|r
+        - Reworked Sync
+            - Now checks version information before syncing.
+            - Will now support larger sized data sets.
+            - Not so spammy now.
         - Rework of invalid zones to support other languages.
         - Also, made seasonal dungeons and raids automatically added to invalid zones.
         - Settings Rework: